hw6/hw6/src/main/java/ru/otus/hw6/repository/genre/GenreRepositoryJpa.java
package ru.otus.hw6.repository.genre;


import jakarta.persistence.EntityManager;
import jakarta.persistence.PersistenceContext;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;
import ru.otus.hw6.domain.Genre;

import java.util.List;

@Component
@RequiredArgsConstructor
public class GenreRepositoryJpa implements GenreRepository {

    @PersistenceContext
    private final EntityManager em;

    @Override
    public List<Genre> getAll() {
        return em.createQuery("select g from Genre g", Genre.class).getResultList();
    }

    @Override
    public Genre getById(long id) {
        return em.find(Genre.class, id);
    }

    @Override
    public long insert(Genre genre) {
        return em.merge(genre).getId();
    }
}
